backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==3.2.3
django-filter==21.1
djangorestframework==3.12.4
djoser==2.3.1
drf-extra-fields==3.7.0
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==20.1.0
hashids==1.3.1
idna==3.10
Markdown==3.7
oauthlib==3.2.2
openpyxl==3.1.5
pillow==11.1.0
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.9.0
python_dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0